add docker files

+++ ticket-service/src/main/java/com/aston/intensive/ticketservice/controller/TestController.java
@@ -13,7 +13,7 @@ public class TestController {
     private String eurekaInstanceId;
 
     @Value("${temporaryVariable}")
-    private Integer temporaryVariable;
+    private Long temporaryVariable;
 
     @GetMapping("/test")
     public String test() {
